feat: ãQNA MAPPER JAVA,XML FIXED

+++ src/main/java/org/esnack24api/esnack24api/customersupport/domain/QNAEntity.java
@@ -19,11 +19,11 @@ import java.time.LocalDateTime;
 public class QNAEntity {
 
     @Id
-    private int qno;
+    private Long qno;
 
-    private int uno;
+    private Long uno;
 
-    private int pno;
+    private Long pno;
 
     private String qtitle;
